requirements.txt
Flask==2.1.0
flask-restful
google-cloud-datastore
